Add project env
